reorganise endpoint files
