+++ mcp_bridge/api_endpoints/openai/__init__.py
@@ -2,12 +2,10 @@ from fastapi import APIRouter
 
 from lmos_openai_types import CreateChatCompletionRequest, CreateCompletionRequest
 
-from openai_clients import (
-    client,
-    completions,
-    chat_completions,
-    streaming_chat_completions,
-)
+from .genericHttpxClient import client
+from .completion import completions
+from .chatCompletion import chat_completions
+from .streamChatCompletion import streaming_chat_completions
 
 from openapi_tags import Tag
