+++ mcp_bridge/api_endpoints/openai/completion.py
