--- mcp_bridge/openai_clients/__init__.py
@@ -1,6 +1,0 @@
-from .genericHttpxClient import client
-from .completion import completions
-from .chatCompletion import chat_completions
-from .streamChatCompletion import streaming_chat_completions
-
-__all__ = ["client", "completions", "chat_completions", "streaming_chat_completions"]
